% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotiff.R
\name{geotiff}
\alias{geotiff}
\title{GeoTIFF}
\usage{
geotiff()
}
\value{
string, either "GDALReadOnlyDataset" or "try-error"
}
\description{
Open GDAL GeoTIFF autotest file.
}
\details{
See mdsumner/gdal/data-raw/  for details on getting this file.
}
\examples{
geotiff()
}

