% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf4.R
\name{hdf4}
\alias{hdf4}
\title{GeoTIFF}
\usage{
hdf4()
}
\value{
string, either "GDALReadOnlyDataset" or "try-error"
}
\description{
Open GDAL GeoTIFF autotest file.
}
\details{
See mdsumner/gdal/data-raw/  for details on getting this file.
}
\examples{
hdf4()
}

